+++ CHANGELOG.md
@@ -6,6 +6,13 @@ File format follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 See cloning and downloading instructions [here](https://www.ivoyager.dev/developers/).
 
+## [v0.0.20] - 2024-12-20
+
+Developed using Godot 4.3.
+
+### Fixed
+* Export breaking reference to EditorInterface outside of EditorPlugin.
+
 ## [v0.0.19] - 2024-12-16
 
 Developed using Godot 4.3.
@@ -84,5 +91,6 @@ Requires plugin [ivoyager_table_reader](https://github.com/ivoyager/ivoyager_tab
 ##
 I, Voyager projects v0.0.16 and earlier used a different core submodule [ivoyager](https://github.com/ivoyager/ivoyager) (now depreciated); see previous changelog [here](https://github.com/ivoyager/ivoyager/blob/master/CHANGELOG.md).
 
+[v0.0.20]: https://github.com/ivoyager/ivoyager_core/compare/v0.0.19...v0.0.20
 [v0.0.19]: https://github.com/ivoyager/ivoyager_core/compare/v0.0.18...v0.0.19
 [v0.0.18]: https://github.com/ivoyager/ivoyager_core/compare/v0.0.17...v0.0.18